pyproject.toml
[project]
name = "protein-language-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.85",
    "einops>=0.8.1",
    "hydra-core>=1.3.2",
    "lightning>=2.5.2",
    "pytest>=8.4.1",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.53.0",
    "wandb>=0.20.1",
]
